fix: export missing consts and types
