--- src/parser/entities/attribute/consts.ts
@@ -1,6 +1,0 @@
-export enum AttributeFlag {
-    INVISIBLE = 1,
-    CONSTANT = 2,
-    REQUIRE_VERIFICATION = 4,
-    PRESET = 8,
-}
