+++ src/parser/entities/attribute/index.ts
@@ -1,3 +1,2 @@
 export type * from './types';
-export * from './consts';
 export * from './parser';
