--- src/parser/entities/types.ts
@@ -1,22 +1,0 @@
-export type * from "./arc";
-export type * from "./attdef";
-export type * from "./circle";
-export type * from "./dimension";
-export type * from "./ellipse";
-export type * from './face'
-export type * from "./hatch";
-export type * from './image'
-export type * from "./insert";
-export type * from "./leader";
-export type * from "./line";
-export type * from "./lwpolyline";
-export type * from "./mtext";
-export type * from "./point";
-export type * from "./polyline";
-export type * from "./ray"
-export type * from "./section";
-export type * from "./solid";
-export type * from "./spline";
-export type * from "./text";
-export type * from "./vertex";
-export type * from "./viewport";
