+++ src/parser/index.ts
@@ -1,9 +1,8 @@
-export * from "./blocks";
-export * from "./types";
+export type * from "./types";
 export * from "./blocks";
 export * from "./entities";
 export * from "./header";
 export * from "./objects";
-export * from "./shared";
 export * from "./tables";
+export * from "./shared";
 export * from './getAcadColor'
