+++ src/parser/objects/dictionary/consts.ts
@@ -1,0 +1,8 @@
+export enum RecordCloneFlag {
+    NOT_APPLICABLE = 0,
+    KEEP_EXISTING = 1,
+    USE_CLONE = 2,
+    XREF_VALUE_NAME = 3,
+    VALUE_NAME = 4,
+    UNMANGLE_NAME = 5,
+}
