+++ src/parser/objects/dictionary/index.ts
@@ -1,0 +1,3 @@
+export type * from './types'
+export * from './consts'
+export * from './parser'
