+++ src/parser/objects/dictionary/parser.ts
@@ -2,8 +2,8 @@ import {
     DXFParserSnippet,
     Identity,
     ToBoolean,
-} from '../shared/parserGenerator';
-import type { CommonDXFObject } from './common';
+} from '@src/parser/shared/parserGenerator';
+import type { DictionaryDXFObject } from './types';
 
 export const DictionarySnippets: DXFParserSnippet[] = [
     {
@@ -44,18 +44,5 @@ export const DictionarySnippets: DXFParserSnippet[] = [
     },
 ];
 
-export enum RecordCloneFlag {
-    NOT_APPLICABLE = 0,
-    KEEP_EXISTING = 1,
-    USE_CLONE = 2,
-    XREF_VALUE_NAME = 3,
-    VALUE_NAME = 4,
-    UNMANGLE_NAME = 5,
-}
 
-export interface DictionaryDXFObject extends CommonDXFObject {
-    subclassMarker: 'AcDbDictionary';
-    isHardOwned?: boolean;
-    recordCloneFlag: RecordCloneFlag;
-    entries: { name: string; objectId?: string }[];
-}
+
