+++ src/parser/objects/layout/consts.ts
@@ -1,0 +1,5 @@
+export enum LayoutControlFlag {
+    PSLTSCALE = 1,
+    LIMCHECK = 2,
+}
+
