+++ src/parser/objects/layout/parser.ts
@@ -1,11 +1,9 @@
-import type { OrthographicType } from '../../consts';
-import type { Point2D, Point3D } from '../../types';
 import {
     DXFParserSnippet,
     Identity,
     PointParser,
-} from '../shared/parserGenerator';
-import { PlotSettingDXFObject, PlotSettingsSnippets } from './plotSettings';
+} from '@src/parser/shared/parserGenerator';
+import { PlotSettingsSnippets } from '../plotSettings';
 
 // Snippet이 code별로 스택에 들어가기 때문에 일부로 역순으로 적음
 export const LayoutSnippets: DXFParserSnippet[] = [
@@ -107,30 +105,3 @@ export const LayoutSnippets: DXFParserSnippet[] = [
     ...PlotSettingsSnippets,
 ];
 
-export enum LayoutControlFlag {
-    PSLTSCALE = 1,
-    LIMCHECK = 2,
-}
-
-export interface LayoutDXFObject
-    extends Omit<PlotSettingDXFObject, 'subclassMarker'> {
-    subclassMarker: 'AcDbLayout';
-    layoutName: string;
-    controlFlag: LayoutControlFlag;
-    tabOrder: number;
-    minLimit: Point2D;
-    maxLimit: Point2D;
-    insertionPoint: Point3D;
-    minExtent: Point3D;
-    maxExtent: Point3D;
-    elevation: number;
-    ucsOrigin: Point3D;
-    ucsXAxis: Point3D;
-    ucsYAxis: Point3D;
-    orthographicType: OrthographicType;
-    paperSpaceTableId: string;
-    viewportId: string;
-    namedUcsId?: string;
-    orthographicUcsId?: string;
-    shadePlotId: string;
-}
