+++ src/parser/objects/shared.ts
@@ -46,10 +46,3 @@ export const CommonObjectSnippets: DXFParserSnippet[] = [
         parser: Identity,
     },
 ];
-
-export interface CommonDXFObject {
-    ownerObjectId: string;
-    ownerDictionaryIdHard: string;
-    ownerDictionaryIdSoft: string;
-    handle: string;
-}
