+++ src/parser/tables/index.ts
@@ -1,11 +1,8 @@
-export type * from './blockRecord'
-export type * from './dimStyle'
-export type * from './layer'
-export type * from './ltype'
-export type * from './style'
-export type * from './vport'
+export * from './blockRecord'
+export * from './dimStyle'
+export * from './layer'
+export * from './ltype'
+export * from './style'
+export * from './vport'
 export type * from './types';
-
-export * from './dimStyle/consts'
-export * from './ltype/consts'
 export * from './parser';
